refactor(layout): move routes constants to lib/constants and add search functionality
